Add files via upload
This is a kaggle dataset hotel_bookings which I utilised to derive the following insights -
They want to understand if this increase is due to seasonality by comparing the total number of guests (monthly) from January 1, 2017 – August 31, 2017 with the same period in 2016. 
I have attached technical design document as well for those who want to practice can download. Thanks